remove duplicate code
Fix the issue #3639
